trying to connect the news data from backend on frontend. Stuck on the useQuery
