+++ package.json
@@ -12,12 +12,14 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "@apollo/client": "^3.8.6",
     "@types/node": "20.6.3",
     "@types/react": "18.2.22",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "10.4.16",
     "axios": "^1.5.0",
     "eslint-config-next": "13.5.2",
+    "graphql": "^16.8.1",
     "next": "13.5.2",
     "postcss": "8.4.30",
     "react": "18.2.0",
